// src/laana-v0.1.3/types/event/wrapper.ts
export interface EventWrapper {
    /**
     * @generated from protobuf field: string eventId = 1;
     */
    eventId: string;
    /**
     * @generated from protobuf field: uint64 time = 2;
     */
    time: bigint;
    /**
     * @generated from protobuf field: string botUin = 3;
     */
    botUin: string;
    /**
     * @generated from protobuf oneof: event
     */
    event: {
        oneofKind: "buddyAddedEvent";
        /**
         * @generated from protobuf field: Laana.BuddyAddedEvent buddyAddedEvent = 101;
         */
        buddyAddedEvent: BuddyAddedEvent;
    } | {
        oneofKind: "buddyMessageWithdrawEvent";
        /**
         * @generated from protobuf field: Laana.BuddyMessageWithdrawEvent buddyMessageWithdrawEvent = 102;
         */
        buddyMessageWithdrawEvent: BuddyMessageWithdrawEvent;
    } | {
        oneofKind: "groupAdminChangeEvent";
        /**
         * @generated from protobuf field: Laana.GroupAdminChangeEvent groupAdminChangeEvent = 201;
         */
        groupAdminChangeEvent: GroupAdminChangeEvent;
    } | {
        oneofKind: "groupEssenceMessageEvent";
        /**
         * @generated from protobuf field: Laana.GroupEssenceMessageEvent groupEssenceMessageEvent = 202;
         */
        groupEssenceMessageEvent: GroupEssenceMessageEvent;
    } | {
        oneofKind: "groupMemberCardChangeEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberCardChangeEvent groupMemberCardChangeEvent = 203;
         */
        groupMemberCardChangeEvent: GroupMemberCardChangeEvent;
    } | {
        oneofKind: "groupMemberIncreaseEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberIncreaseEvent groupMemberIncreaseEvent = 204;
         */
        groupMemberIncreaseEvent: GroupMemberIncreaseEvent;
    } | {
        oneofKind: "groupMemberDecreaseEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberDecreaseEvent groupMemberDecreaseEvent = 205;
         */
        groupMemberDecreaseEvent: GroupMemberDecreaseEvent;
    } | {
        oneofKind: "groupMemberTitleChangeEvent";
        /**
         * @generated from protobuf field: Laana.GroupMemberTitleChangeEvent groupMemberTitleChangeEvent = 206;
         */
        groupMemberTitleChangeEvent: GroupMemberTitleChangeEvent;
    } | {
        oneofKind: "groupMessageWithdrawEvent";
        /**
         * @generated from protobuf field: Laana.GroupMessageWithdrawEvent groupMessageWithdrawEvent = 207;
         */
        groupMessageWithdrawEvent: GroupMessageWithdrawEvent;
    } | {
        oneofKind: "groupMuteMemberEvent";
        /**
         * @generated from protobuf field: Laana.GroupShutUpMemberEvent groupMuteMemberEvent = 208;
         */
        groupMuteMemberEvent: GroupShutUpMemberEvent;
    } | {
        oneofKind: "groupMuteAllEvent";
        /**
         * @generated from protobuf field: Laana.GroupShutUpAllEvent groupMuteAllEvent = 209;
         */
        groupMuteAllEvent: GroupShutUpAllEvent;
    } | {
        oneofKind: "groupUnmuteMemberEvent";
        /**
         * @generated from protobuf field: Laana.GroupLiftShutUpMemberEvent groupUnmuteMemberEvent = 210;
         */
        groupUnmuteMemberEvent: GroupLiftShutUpMemberEvent;
    } | {
        oneofKind: "groupUnmuteAllEvent";
        /**
         * @generated from protobuf field: Laana.GroupLiftShutUpAllEvent groupUnmuteAllEvent = 211;
         */
        groupUnmuteAllEvent: GroupLiftShutUpAllEvent;
    } | {
        oneofKind: "groupEmojiLikeEvent";
        /**
         * @generated from protobuf field: Laana.GroupEmojiLikeEvent groupEmojiLikeEvent = 212;
         */
        groupEmojiLikeEvent: GroupEmojiLikeEvent;
    } | {
        oneofKind: "someoneRequestToAddBuddyEvent";
        /**
         * @generated from protobuf field: Laana.SomeoneRequestToAddBuddyEvent someoneRequestToAddBuddyEvent = 301;
         */
        someoneRequestToAddBuddyEvent: SomeoneRequestToAddBuddyEvent;
    } | {
        oneofKind: "someoneRequestToJoinGroupEvent";
        /**
         * @generated from protobuf field: Laana.SomeoneRequestToJoinGroupEvent someoneRequestToJoinGroupEvent = 302;
         */
        someoneRequestToJoinGroupEvent: SomeoneRequestToJoinGroupEvent;
    } | {
        oneofKind: "selfInvitedToGroupEvent";
        /**
         * @generated from protobuf field: Laana.SelfInvitedToGroupEvent selfInvitedToGroupEvent = 303;
         */
        selfInvitedToGroupEvent: SelfInvitedToGroupEvent;
    } | {
        oneofKind: undefined;
    };
}
// ...
export const EventWrapper = new EventWrapper$Type();